requirements.txt
wheel
cryptography
mysql-connector-python